// lib/catalog/vendor-privacy-types.ts
export type LockedVendorContactAvailability = {
  whatsappPrimaryAvailable: boolean;
  whatsappSecondaryAvailable: boolean;
  telegramAvailable: boolean;
  instagramAvailable: boolean;
  phoneAvailable: boolean;
  googleMapsAvailable: boolean;
  twoGisAvailable: boolean;
  yandexMapsAvailable: boolean;
};